ada yg diubah dikit di migration sm seeder

+++ database/migrations/2024_11_15_011423_remove_status_persetujuan_from_ruangs_table.php
@@ -4,25 +4,29 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class RemoveStatusPersetujuanFromRuangsTable extends Migration
 {
     /**
      * Run the migrations.
      */
     public function up()
-{
-    Schema::table('ruangs', function (Blueprint $table) {
-        $table->dropColumn('status_persetujuan');
-    });
-}
+    {
+        if (Schema::hasColumn('ruang', 'status_persetujuan')) {
+            Schema::table('ruang', function (Blueprint $table) {
+                $table->dropColumn('status_persetujuan');
+            });
+        }
+    }
 
     /**
      * Reverse the migrations.
      */
     public function down()
-{
-    Schema::table('ruangs', function (Blueprint $table) {
-        $table->string('status_persetujuan')->default('Belum Disetujui');
-    });
+    {
+        Schema::table('ruang', function (Blueprint $table) {
+            if (!Schema::hasColumn('ruang', 'status_persetujuan')) {
+                $table->string('status_persetujuan')->default('Belum Disetujui');
+            }
+        });
+    }
 }
-};
